Created mixins
